requirements.txt
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==2.0.2
Flask-Cors==3.0.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
six==1.16.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
